# release-2.1: opt: Improve performance of IndexJoin planning