--- a/pkg/sql/opt/xform/testdata/rules/stats
+++ b/pkg/sql/opt/xform/testdata/rules/stats
@@ -35,10 +35,9 @@ rulestats
 SELECT * FROM abc JOIN xyz ON a=x
 ----
 Normalization rules applied 0 times.
-Exploration rules applied 56 times, added 49 expressions.
+Exploration rules applied 8 times, added 5 expressions.
 Top exploration rules:
-  CommuteJoin              applied 18 times, added 9  expressions.
-  GenerateMergeJoins       applied 18 times, added 18 expressions.
-  PushJoinThroughIndexJoin applied 12 times, added 12 expressions.
-  GenerateLookupJoin       applied 6  times, added 6  expressions.
-  GenerateIndexScans       applied 2  times, added 4  expressions.
+  CommuteJoin        applied 2 times, added 1 expressions.
+  GenerateMergeJoins applied 2 times, added 2 expressions.
+  GenerateLookupJoin applied 2 times, added 2 expressions.
+  GenerateIndexScans applied 2 times, added 0 expressions.