… pkg/sql/opt/xform/index_scan_builder.go
… pkg/sql/opt/xform/rules/select.opt
… pkg/sql/opt/xform/testdata/external/activerecord
… pkg/sql/opt/xform/testdata/external/nova
… pkg/sql/opt/xform/testdata/rules/join
… pkg/sql/opt/xform/testdata/rules/scan
… pkg/sql/opt/xform/testdata/rules/select